requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==3.7.1
async-timeout==5.0.1
attrs==25.3.0
boto3==1.39.3
botocore==1.39.3
click==8.1.8
dotenv==0.9.9
exceptiongroup==1.3.0
fastapi==0.104.1
frozenlist==1.6.0
h11==0.16.0
idna==3.10
jmespath==1.0.1
livekit-api==1.0.2
livekit-protocol==1.0.2
mangum==0.17.0
multidict==6.4.3
propcache==0.3.1
protobuf==6.30.2
pydantic==2.4.2
pydantic_core==2.10.1
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
s3transfer==0.13.0
six==1.17.0
sniffio==1.3.1
starlette==0.27.0
types-protobuf==6.30.2.20250506
typing_extensions==4.13.2
urllib3==1.26.20
uvicorn==0.23.2
yarl==1.20.0
